database changes

+++ wp-content/plugins/image-slider-widget/inc/functions/ewic-functions.php
@@ -82,8 +82,8 @@ function ewic_img_remove() {
 		}
 		
 		else {
-
-			update_post_meta( $_POST['pstid'], 'ewic_meta_select_images', '' );
+			$thepstid = sanitize_text_field( $_POST['pstid'] );
+			update_post_meta( $thepstid, 'ewic_meta_select_images', '' );
 			
 			echo '1';
 			
@@ -142,13 +142,15 @@ function ewic_ajax_autoupdt() {
 	
 	check_ajax_referer( 'easywic-lite-nonce', 'security' );
 	
-	if ( !isset( $_POST['cmd'] ) ) {
+	$thecmd = sanitize_text_field( $_POST['cmd'] );
+	
+	if ( !isset( $thecmd ) ) {
 		echo '0';
 		wp_die();
 		}
 		
 		else {
-			update_option( "ewic-settings-automatic_update", $_POST['cmd'] );	
+			update_option( "ewic-settings-automatic_update", $thecmd );	
 			echo '1';	
 			wp_die();
 			}
@@ -204,7 +206,7 @@ function ewic_share() {
 <div style="position:relative; margin-top:6px;">
 <ul class='ewic-social' id='ewic-cssanime'>
 <li class='ewic-facebook'>
-<a onclick="window.open('http://www.facebook.com/sharer.php?s=100&amp;p[title]=Check out the Best Image Slider Wordpress Plugin&amp;p[summary]=Best Image Slider Wordpress Plugin is powerful plugin to create image slider in minutes&amp;p[url]=http://demo.ghozylab.com/plugins/easy-image-slider-plugin/&amp;p[images][0]=http://content.ghozylab.com/wp-content/uploads/2014/11/easy-slider-widget-320-200.png', 'sharer', 'toolbar=0,status=0,width=548,height=325');" href="javascript: void(0)" title="Share"><strong>Facebook</strong></a>
+<a onclick="window.open('http://www.facebook.com/sharer.php?s=100&amp;p[title]=Check out the Best Image Slider Wordpress Plugin&amp;p[summary]=Best Image Slider Wordpress Plugin is powerful plugin to create image slider in minutes&amp;p[url]=http://demo.ghozylab.com/plugins/easy-image-slider-plugin/&amp;p[images][0]=<?php echo EWIC_URL . '/images/assets/easy-slider-widget-320-200.png'; ?>', 'sharer', 'toolbar=0,status=0,width=548,height=325');" href="javascript: void(0)" title="Share"><strong>Facebook</strong></a>
 </li>
 <li class='ewic-twitter'>
 <a onclick="window.open('https://twitter.com/share?text=Best Wordpress Image Slider Plugin &url=http://demo.ghozylab.com/plugins/easy-image-slider-plugin/', 'sharer', 'toolbar=0,status=0,width=548,height=325');" title="Twitter" class="circle"><strong>Twitter</strong></a>
@@ -213,7 +215,7 @@ function ewic_share() {
 <a onclick="window.open('https://plus.google.com/share?url=http://demo.ghozylab.com/plugins/easy-image-slider-plugin/','','width=415,height=450');"><strong>Google+</strong></a>
 </li>
 <li class='ewic-pinterest'>
-<a onclick="window.open('http://pinterest.com/pin/create/button/?url=http://demo.ghozylab.com/plugins/easy-image-slider-plugin/;media=http://content.ghozylab.com/wp-content/uploads/2014/11/easy-slider-widget-320-200.png;description=Best Image Slider Wordpress Plugin','','width=600,height=300');"><strong>Pinterest</strong></a>
+<a onclick="window.open('http://pinterest.com/pin/create/button/?url=http://demo.ghozylab.com/plugins/easy-image-slider-plugin/;media=<?php echo EWIC_URL . '/images/assets/easy-slider-widget-320-200.png'; ?>;description=Best Image Slider Wordpress Plugin','','width=600,height=300');"><strong>Pinterest</strong></a>
 </li>
 </ul>
 </div>
@@ -284,7 +286,8 @@ function ewic_enable_auto_update() {
 		
 		else {
 			if ( $_POST['cmd'] == 'active' ){
-				update_option( "ewic-settings-automatic_update", $_POST['cmd'] );
+				$thecmd = sanitize_text_field( $_POST['cmd'] );
+				update_option( "ewic-settings-automatic_update", $thecmd );
 				echo '1';				
 				wp_die();
 				}
@@ -457,7 +460,7 @@ function ewic_get_aff_data() {
 			// listen for aff button to be clicked
 			if( isset( $_POST['eml'] ) ) {
 				
-				$affemail = $_POST['eml'];
+				$affemail = sanitize_email( $_POST['eml'] );
 				
 				$api_params = array(
 					'ghozy_action' => 'get_aff_data',
@@ -555,22 +558,30 @@ function ewic_timthumb_check() {
 	
 	check_ajax_referer( 'ewic-tt-nonce', 'security' );
 	
-	$file = $_POST['turl'];
+	if ( is_admin() && current_user_can( 'manage_options' ) ) {
 	
-	$checknow = @fopen( $file, 'w' );
-	
-	if( $checknow ){
-			
-		fclose( $checknow );
-		unlink( $file );
+		$file = $_POST['turl'];
 		
-			echo '1';
-
-		} else {
+		$checknow = @fopen( $file, 'w' );
+		
+		if( $checknow ){
+				
+			fclose( $checknow );
+			unlink( $file );
 			
-			wp_die();
+				echo '1';
+	
+			} else {
+				
+				wp_die();
+				
+			}
 			
-		}
+	} else {
+		
+		wp_die();
+			
+	}
 		
 	wp_die();
